Remove all vestiges of Game Feature plugin
Unfortunately, game feature plugins don't seem to be baked enough to use for anything other than content, and aren't something I'm willing to take on at the moment. The separability guarantees are nice, but broadly, useless to me at the moment. That and it adds yet another config file, this one stored as a binary asset file. Yes, there's a tool for doing merges of data assets, but really, guys? Really? That's a config file. You have those!