Adjust, plan and prep for the thistle service
Thistle looks like it's shaping up to be a system for swapping behavior trees, store behavior tree assets, connect behavior trees, and put the services we need to provide asynchronous or network task resolution to behavior trees.  Down the line, we should consider exposing it through statetrees, allowing a state tree to act as the definition of which pawns should be controlled how. At the moment, state trees don't necessarily feel very baked to me.

It's likely that the BULK of thistle will boil down to updating blackboards, swapping behavior trees, and compartmentalizing AI.