UE AI does not lend itself towards SoC or composition by default
but I think that's actually more the tutorial and the normal use cases than anything else. by and at large, a lot of this boils down to:

Side-effects are the norm, not the exception.
Network synchro is not TRULY considered in the design.
BP designs tend towards conflation due to the graph nature.